+++ packages/kit/test/apps/basics/src/routes/nested-layout/error.svelte
@@ -1,0 +1,7 @@
+<script context="module">
+	export function load() {
+		return { status: 500, error: 'Error' };
+	}
+</script>
+
+<h1>Not shown</h1>