+++ packages/adapter-node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-node
 
+## 1.0.0-next.46
+
+### Patch Changes
+
+- Clear output directory before adapting ([#2388](https://github.com/sveltejs/kit/pull/2388))
+
+## 1.0.0-next.45
+
+### Patch Changes
+
+- [chore] export package.json from adapters ([#2351](https://github.com/sveltejs/kit/pull/2351))
+
 ## 1.0.0-next.44
 
 ### Patch Changes